requirements.txt
python-dotenv
feedparser>=6.0.10
sgmllib3k
langdetect
beautifulsoup4
markdownify
requests
rich
pyyaml